Add files via upload
Presentation giving an overview of the project, breaking down concepts of why it was created and for what uses. We were given goals to use a software we had no experience using, and took state derived data displaying the street overview of the town of New Paltz. We used this data to take two reference points, one being the current firehouse location, and an additional location for a foreseeable firehouse. The idea is to use this software to predict where would be the most efficient location for an additional firehouse, to evenly disperse routing times to destinations around New Paltz.